style: add setting navigation
